chimera-engine/vital-POC/requirements.txt
openai
faiss-cpu
python-dotenv
streamlit
